Правки после первого ревью
refactor: правки после первого ревью

+++ src/main/java/ru/yandex/practicum/filmorate/controllers/DirectorController.java
@@ -26,7 +26,7 @@ public class DirectorController {
     }
 
     @GetMapping("/{id}")
-    public Director get(@PathVariable int id) {
+    public Director get(@PathVariable Long id) {
         return directorService.get(id);
     }
 
@@ -36,7 +36,7 @@ public class DirectorController {
     }
 
     @DeleteMapping("/{id}")
-    public void delete(@PathVariable int id) {
+    public void delete(@PathVariable Long id) {
         directorService.delete(id);
     }
 }
