+++ src/main/java/ru/yandex/practicum/filmorate/service/MPAARatingService.java
@@ -20,7 +20,7 @@ public class MPAARatingService {
         return mpaaRatingStorage.getAll();
     }
 
-    public MPAAFilmRating getById(int mpaaId) {
+    public MPAAFilmRating getById(Long mpaaId) {
         MPAAFilmRating mpa = mpaaRatingStorage.getById(mpaaId).orElseThrow(
                 () -> new MPAARatingNotFoundException("Возрастного рейтина с id: " + mpaaId + " не существует.")
         );
