+++ src/main/java/ru/yandex/practicum/filmorate/storage/likes/LikesDbStorage.java
@@ -13,13 +13,13 @@ public class LikesDbStorage implements LikesStorage {
     private final JdbcTemplate jdbcTemplate;
 
     @Override
-    public void addLike(int filmId, int userId) {
+    public void addLike(Long filmId, Long userId) {
         final String sqlQuery = "INSERT INTO liked (FILM_ID, USER_ID) VALUES (?, ?) ON CONFLICT DO NOTHING ";
         jdbcTemplate.update(sqlQuery, filmId, userId);
     }
 
     @Override
-    public void deleteLike(int filmId, int userId) {
+    public void deleteLike(Long filmId, Long userId) {
         final String sqlQuery = "DELETE FROM liked WHERE FILM_ID = ? AND USER_ID = ?";
         int status = jdbcTemplate.update(sqlQuery, filmId, userId);
         if (status != 1) {
@@ -28,8 +28,8 @@ public class LikesDbStorage implements LikesStorage {
     }
 
     @Override
-    public List<Integer> getListOfLikes(int userId) {
+    public List<Long> getListOfLikes(Long userId) {
         final String sqlQuery = "SELECT FILM_ID FROM liked WHERE USER_ID = ?;";
-        return jdbcTemplate.queryForList(sqlQuery, Integer.class, userId);
+        return jdbcTemplate.queryForList(sqlQuery, Long.class, userId);
     }
 }
