+++ src/main/java/ru/yandex/practicum/filmorate/storage/likes/LikesStorage.java
@@ -1,13 +1,12 @@
 package ru.yandex.practicum.filmorate.storage.likes;
 
-
 import java.util.List;
 
 public interface LikesStorage {
 
-    void addLike(int filmId, int userId);
+    void addLike(Long filmId, Long userId);
 
-    void deleteLike(int filmId, int userId);
+    void deleteLike(Long filmId, Long userId);
 
-    List<Integer> getListOfLikes(int userId);
+    List<Long> getListOfLikes(Long userId);
 }
